starsh/requirements.txt
Django == 4.1.*
wagtail == 4.1.*
PyYAML
Markdown
Pygments
djangorestframework
